refactor finished for view pages
