+++ lib/src/components/pages/view/activities.dart
@@ -8,6 +8,7 @@ import 'package:wui_builder/vhtml.dart';
 import 'package:built_collection/built_collection.dart';
 
 import '../../core/nav.dart';
+import '../../core/pageRepeats.dart';
 import '../../../model/activity.dart';
 import '../../../model/user.dart';
 import '../../../state/app.dart';
