+++ lib/src/components/pages/view/members.dart
@@ -6,6 +6,7 @@ import 'package:wui_builder/vhtml.dart';
 import 'package:built_collection/built_collection.dart';
 
 import '../../core/nav.dart';
+import '../../core/pageRepeats.dart';
 import '../../../constants.dart';
 import '../../../model/user.dart';
 import '../../../model/activity.dart';
@@ -47,6 +48,34 @@ class ViewMembers extends Component<ViewMembersProps, ViewMembersState> {
   VNode emailInputNode;
   VNode passwordInputNode;
 
+  @override
+  VNode render() => new VDivElement()
+    ..children = [
+      new Nav(new NavProps()
+        ..actions = props.actions
+        ..user = props.user),
+      new VDivElement()
+        ..className = 'container'
+        ..children = [
+          new VDivElement()
+            ..className = 'columns is-mobile margin-top is-centered'
+            ..children = [
+              new VDivElement()
+                ..className = 'column is-four-fifths'
+                ..children = [
+                  new VDivElement()
+                    ..className = 'box is-4'
+                    ..children = [
+                      _renderHeader(),
+                      new VTableElement()
+                        ..className = 'table is-striped is-fullwidth'
+                        ..children = _createRows(),
+                    ],
+                ],
+            ],
+        ],
+    ];
+
   /// [_createRows] Scaling function to make rows based on amount of information available
   List<VNode> _createRows() {
     List<User> users;
@@ -57,7 +86,7 @@ class ViewMembers extends Component<ViewMembersProps, ViewMembersState> {
     } else {
       users = state.found;
     }
-    nodeList.addAll(_titleRow());
+    nodeList.addAll(titleRow(title));
     //merge sort function by last name
     users = _sort(users, 0, users.length - 1);
     for (User user in users) {
@@ -103,14 +132,6 @@ class ViewMembers extends Component<ViewMembersProps, ViewMembersState> {
     return nodeList;
   }
 
-  _onUserClick(String uid) {
-    history.push(Routes.generateEditMemberURL(uid));
-  }
-
-  _onActClick(String uid) {
-    history.push(Routes.generateActivitySignUpURL(uid));
-  }
-
   /// [sort] Merge sort by last name of user
   List<User> _sort(List<User> users, int left, int right) {
     if (left < right) {
@@ -152,48 +173,6 @@ class ViewMembers extends Component<ViewMembersProps, ViewMembersState> {
     return users;
   }
 
-  /// [_titleRow] helper function to create the title row
-  List<VNode> _titleRow() {
-    List<VNode> nodeList = new List();
-    for (String title in title) {
-      nodeList.add(
-        new VTableCellElement()
-          ..className = 'title is-5'
-          ..text = title,
-      );
-    }
-    return nodeList;
-  }
-
-  @override
-  VNode render() => new VDivElement()
-    ..children = [
-      new Nav(new NavProps()
-        ..actions = props.actions
-        ..user = props.user),
-      new VDivElement()
-        ..className = 'container'
-        ..children = [
-          new VDivElement()
-            ..className = 'columns is-mobile margin-top is-centered'
-            ..children = [
-              new VDivElement()
-                ..className = 'column is-four-fifths'
-                ..children = [
-                  new VDivElement()
-                    ..className = 'box is-4'
-                    ..children = [
-                      _renderHeader(),
-                      new VTableElement()
-                        ..className = 'table is-striped is-fullwidth'
-                        ..children = _createRows(),
-                    ],
-                ],
-            ],
-          // _modView(),
-        ],
-    ];
-
   ///[_renderHeader] makes the title bar of the viewMembers page
   _renderHeader() => new VDivElement()
     ..className = 'columns is-mobile'
@@ -208,76 +187,9 @@ class ViewMembers extends Component<ViewMembersProps, ViewMembersState> {
             ..className = 'subtitle is-7'
             ..text = "as of: ${DateTime.now().month}/${DateTime.now().day}/${DateTime.now().year}",
         ],
-      _renderSearch(),
-      _renderExport(),
-      _renderRefresh(),
-    ];
-
-  ///[_renderSearch] adds the seach layout to the tile bar
-  _renderSearch() => new VDivElement()
-    ..className = 'column is-narrow'
-    ..children = [
-      new VDivElement()
-        ..className = 'field'
-        ..children = [
-          new VParagraphElement()
-            ..className = 'control has-icons-left'
-            ..children = [
-              new VInputElement()
-                ..className = 'input'
-                ..placeholder = 'Search'
-                ..id = 'Search'
-                ..onKeyUp = _searchListener
-                ..type = 'text',
-              new VSpanElement()
-                ..className = 'icon is-left'
-                ..children = [new Vi()..className = 'fas fa-search'],
-            ],
-        ],
-    ];
-
-  _renderRefresh() => new VDivElement()
-    ..className = 'column is-narrow'
-    ..children = [
-      new VDivElement()
-        ..className = 'field'
-        ..children = [
-          new VDivElement()
-            ..className = 'control'
-            ..children = [
-              new VParagraphElement()
-                ..className = 'button is-rounded'
-                ..onClick = _onRefreshClick
-                ..children = [
-                  new VSpanElement()
-                    ..className = 'icon'
-                    ..children = [new Vi()..className = 'fas fa-sync-alt'],
-                  new VSpanElement()..text = 'Refresh',
-                ],
-            ],
-        ],
-    ];
-
-  _renderExport() => new VDivElement()
-    ..className = 'column is-narrow'
-    ..children = [
-      new VDivElement()
-        ..className = 'field'
-        ..children = [
-          new VDivElement()
-            ..className = 'control'
-            ..children = [
-              new VParagraphElement()
-                ..className = 'button is-rounded'
-                ..onClick = _onExportCsvClick
-                ..children = [
-                  new VSpanElement()
-                    ..className = 'icon'
-                    ..children = [new Vi()..className = 'fas fa-file-csv'],
-                  new VSpanElement()..text = 'Export',
-                ],
-            ],
-        ],
+      renderSearch(_searchListener),
+      renderExport(_onExportCsvClick),
+      renderRefresh(_onRefreshClick),
     ];
 
   _searchListener(_) {
@@ -356,5 +268,13 @@ class ViewMembers extends Component<ViewMembersProps, ViewMembersState> {
 
   _onRefreshClick(_) {
     props.actions.server.fetchAllMembers();
+  }
+
+  _onUserClick(String uid) {
+    history.push(Routes.generateEditMemberURL(uid));
+  }
+
+  _onActClick(String uid) {
+    history.push(Routes.generateActivitySignUpURL(uid));
   }
 }
